added check on external dive in backup

+++ scripts/backupscript.sh
@@ -1,7 +1,11 @@
 notify-send " Starting Backup"
-sudo rsync -avAXS --delete --info=progress2 --exclude={"/dev/*","/proc/*","/sys/*","/tmp/*","/run/*","/mnt/*","/media/*","/lost+found","/home/*/.thumbnails/*","/home/*/.cache/mozilla/*","/home/*/.cache/chromium/*","/home/*/.local/share/Trash/*"} / /media/Seagate/backup
-if [ $? != 0 ]; then
+if [ -d "/media/Seagate" ]; then
+  sudo rsync -avAXS --delete --info=progress2 --exclude={"/dev/*","/proc/*","/sys/*","/tmp/*","/run/*","/mnt/*","/media/*","/lost+found","/home/*/.thumbnails/*","/home/*/.cache/mozilla/*","/home/*/.cache/chromium/*","/home/*/.local/share/Trash/*"} / /media/Seagate/backup
+  if [ $? != 0 ]; then
 	notify-send "Backup unsucsessful!"
-else
+  else
 	notify-send "Backup successful!"
+  fi
+else
+  notify-send "External drive is not mounted"
 fi
